requirements/production.txt
-r base.txt
# elasticsearch==2.x.x chosen for compatibility with t2.micro.elasticsearch and t2.small.elasticsearch
# instance types on AWS (Elasticsearch 2.3). Adjust for your deployment as needed.
elasticsearch==2.4.1
# Additional dependencies for Heroku deployment
dj-database-url==0.4.1
mysqlclient==1.3.13
uwsgi==2.0.17.1
psycopg2==2.7.3.1
whitenoise==3.2.2
boto==2.45.0
django-storages==1.6.5
# For retrieving credentials and signing requests to Elasticsearch
botocore==1.12.33
aws-requests-auth==0.4.0
django-redis==4.8.0
django_cache_url==2.0.0
# For copying initial media to S3 bucket
awscli==1.16.43
